README.md
The pressing issue of food waste motivates us young students who have a sight of this critical societal need. As we observed, we could see the problem lies in the inefficient distribution of food from those who have it in excess amount to those in need. Current systems often lack coordination and real-time tracking, leading to logistical challenges, food spoilage, and inefficiencies in matching donors with recipients.
 So we have developed a database system that streamlines the process of food donation, collection, and distribution through real-time tracking, automated matching algorithms, and user-friendly interfaces. Our system will provide a centralized platform where donors can input details about their excess food, which will be matched with nearby recipients in need, ensuring timely and efficient distribution while minimizing wastage. The significance of your project: This project holds significant societal impact as it addresses both environmental concerns related to food wastage and humanitarian issues related to hunger. However, it may face challenges such as scalability and funding for expansion. Future work could focus on enhancing scalability, integrating with existing food distribution networks, and expanding outreach to more communities.